gpt/cli/requirements.txt
typer>=0.24.0
httpx>=0.28.1
rich>=14.3.3
python-dotenv>=1.2.1
pydantic>=2.12.5
tabulate>=0.9.0